BG services update, bugfixes

+++ app/core/FileManager.php
@@ -154,7 +154,11 @@ class FileManager {
             }
 
             if(!$root) {
-                return rmdir($dirPath) && $result;
+                if(self::folderExists($dirPath)) {
+                    return rmdir($dirPath) && $result;
+                } else {
+                    return $result;
+                }
             } else {
                 return $result;
             }
